[Ethernet] Split handling WiFi/Eth event processing
